test: pass architecture parameter [ci skip]
execution permissions to bash scripts

chmod a+x test-lambda-maven.sh
chmod a+x test-lambda-maven-graalvm.sh

add echo statements in bash scripts

uncomment statements in bash scripts

+++ test-lambda-graalvm.sh
@@ -1,14 +1,21 @@
 #!/bin/bash
 EXIT_STATUS=0
 
+architecture="$1"
+if [ "$architecture" != "arm" ] && [ "$architecture" != "x86" ]; then
+    echo "First parameter is not equal to 'arm' or 'x86'"
+    exit 1
+fi
+
 rm -rf starter-cli/temp
 
 ############
 # | BUILD  | TYPE     | RUNTIME  | FEATURES
-# | GRADLE | APP      | PROVIDED | jackson-databind,aws-lambda,aws-cdk,amazon-api-gateway-http,graalvm
+# | GRADLE | APP      | PROVIDED | jackson-databind,aws-lambda,aws-cdk,amazon-api-gateway-http,graalvm,$architecture
 ############
 
-./gradlew micronaut-cli:run --args="create-app -b gradle -f \"jackson-databind,aws-lambda,aws-cdk,amazon-api-gateway-http,graalvm\" temp" || EXIT_STATUS=$?
+echo "create-app -b gradle -f \"jackson-databind,aws-lambda,aws-cdk,amazon-api-gateway-http,graalvm,$architecture\" temp"
+./gradlew micronaut-cli:run --args="create-app -b gradle -f \"jackson-databind,aws-lambda,aws-cdk,amazon-api-gateway-http,graalvm,$architecture\" temp" || EXIT_STATUS=$?
 
 if [ $EXIT_STATUS -ne 0 ]; then  
   exit $EXIT_STATUS
@@ -19,7 +26,7 @@ cd starter-cli/temp
 ./test-lambda.sh  || EXIT_STATUS=$? 
 
 if [ $EXIT_STATUS -ne 0 ]; then
-  echo "❌ FAILED | GRADLE | APP | PROVIDED    | jackson-databind,aws-lambda,aws-cdk,amazon-api-gateway-http,graalvm"
+  echo "❌ FAILED | GRADLE | APP | PROVIDED    | jackson-databind,aws-lambda,aws-cdk,amazon-api-gateway-http,graalvm,$architecture"
   exit $EXIT_STATUS
 fi
 
@@ -30,10 +37,10 @@ rm -rf starter-cli/temp
 
 ############
 # | BUILD  | TYPE     | RUNTIME  | FEATURES
-# | GRADLE | FUNCTION | PROVIDED | aws-lambda,aws-cdk,aws-lambda-function-url,graalvm
+# | GRADLE | FUNCTION | PROVIDED | aws-lambda,aws-cdk,aws-lambda-function-url,graalvm,$architecture
 ############
-
-./gradlew micronaut-cli:run --args="create-function-app -b gradle -f \"aws-lambda,aws-cdk,aws-lambda-function-url,graalvm\" temp" || EXIT_STATUS=$?
+echo "create-function-app -b gradle -f \"aws-lambda,aws-cdk,aws-lambda-function-url,graalvm,$architecture\" temp"
+./gradlew micronaut-cli:run --args="create-function-app -b gradle -f \"aws-lambda,aws-cdk,aws-lambda-function-url,graalvm,$architecture\" temp" || EXIT_STATUS=$?
 
 if [ $EXIT_STATUS -ne 0 ]; then  
   exit $EXIT_STATUS
@@ -44,7 +51,7 @@ cd starter-cli/temp
 ./test-lambda.sh  || EXIT_STATUS=$? 
 
 if [ $EXIT_STATUS -ne 0 ]; then
-  echo "❌ FAILED | GRADLE | FUNCTION | PROVIDED    | aws-lambda,aws-cdk,aws-lambda-function-url,graalvm"
+  echo "❌ FAILED | GRADLE | FUNCTION | PROVIDED    | aws-lambda,aws-cdk,aws-lambda-function-url,graalvm,$architecture"
   exit $EXIT_STATUS
 fi
 
@@ -55,10 +62,10 @@ rm -rf starter-cli/temp
 
 ############
 # | BUILD  | TYPE | RUNTIME  | FEATURES
-# | GRADLE | APP  | PROVIDED | aws-lambda,aws-cdk,graalvm,amazon-api-gateway
+# | GRADLE | APP  | PROVIDED | aws-lambda,aws-cdk,graalvm,amazon-api-gateway,$architecture
 ############
-
-./gradlew micronaut-cli:run --args="create-app -b gradle -f \"aws-lambda,aws-cdk,graalvm,amazon-api-gateway\" temp" || EXIT_STATUS=$?
+echo "create-app -b gradle -f \"aws-lambda,aws-cdk,graalvm,amazon-api-gateway,$architecture\" temp"
+./gradlew micronaut-cli:run --args="create-app -b gradle -f \"aws-lambda,aws-cdk,graalvm,amazon-api-gateway,$architecture\" temp" || EXIT_STATUS=$?
 
 if [ $EXIT_STATUS -ne 0 ]; then
   exit $EXIT_STATUS
@@ -69,7 +76,7 @@ cd starter-cli/temp
 ./test-lambda.sh  || EXIT_STATUS=$? 
 
 if [ $EXIT_STATUS -ne 0 ]; then
-  echo "❌ FAILED | GRADLE  | APP  | PROVIDED | aws-lambda,aws-cdk,graalvm,amazon-api-gateway"
+  echo "❌ FAILED | GRADLE  | APP  | PROVIDED | aws-lambda,aws-cdk,graalvm,amazon-api-gateway,$architecture"
   exit $EXIT_STATUS
 fi
 
@@ -80,10 +87,10 @@ rm -rf starter-cli/temp
 
 ############
 # | BUILD  | TYPE | RUNTIME  | FEATURES
-# | GRADLE | APP  | PROVIDED | aws-lambda,aws-cdk,graalvm,amazon-api-gateway-http
+# | GRADLE | APP  | PROVIDED | aws-lambda,aws-cdk,graalvm,amazon-api-gateway-http,$architecture
 ############
-
-./gradlew micronaut-cli:run --args="create-app -b gradle -f \"aws-lambda,aws-cdk,graalvm,amazon-api-gateway-http\" temp" || EXIT_STATUS=$?
+echo "create-app -b gradle -f \"aws-lambda,aws-cdk,graalvm,amazon-api-gateway-http,$architecture\" temp"
+./gradlew micronaut-cli:run --args="create-app -b gradle -f \"aws-lambda,aws-cdk,graalvm,amazon-api-gateway-http,$architecture\" temp" || EXIT_STATUS=$?
 
 if [ $EXIT_STATUS -ne 0 ]; then
   exit $EXIT_STATUS
@@ -94,7 +101,7 @@ cd starter-cli/temp
 ./test-lambda.sh  || EXIT_STATUS=$? 
 
 if [ $EXIT_STATUS -ne 0 ]; then
-  echo "❌ FAILED | GRADLE  | APP  | PROVIDED | aws-lambda,aws-cdk,graalvm,amazon-api-gateway-http"
+  echo "❌ FAILED | GRADLE  | APP  | PROVIDED | aws-lambda,aws-cdk,graalvm,amazon-api-gateway-http,$architecture"
   exit $EXIT_STATUS
 fi
 
@@ -106,10 +113,10 @@ rm -rf starter-cli/temp
 
 ############
 # | BUILD  | TYPE     | RUNTIME  | FEATURES
-# | GRADLE | FUNCTION | PROVIDED | aws-lambda,aws-cdk,graalvm,amazon-api-gateway
+# | GRADLE | FUNCTION | PROVIDED | aws-lambda,aws-cdk,graalvm,amazon-api-gateway,$architecture
 ############
-
-./gradlew micronaut-cli:run --args="create-function-app -b gradle -f \"aws-lambda,aws-cdk,graalvm,amazon-api-gateway\" temp" || EXIT_STATUS=$?
+echo "create-function-app -b gradle -f \"aws-lambda,aws-cdk,graalvm,amazon-api-gateway,$architecture\" temp"
+./gradlew micronaut-cli:run --args="create-function-app -b gradle -f \"aws-lambda,aws-cdk,graalvm,amazon-api-gateway,$architecture\" temp" || EXIT_STATUS=$?
 
 if [ $EXIT_STATUS -ne 0 ]; then
   exit $EXIT_STATUS
@@ -120,7 +127,7 @@ cd starter-cli/temp
 ./test-lambda.sh  || EXIT_STATUS=$? 
 
 if [ $EXIT_STATUS -ne 0 ]; then
-  echo "❌ FAILED | GRADLE  | FUNCTION  | PROVIDED | aws-lambda,aws-cdk,graalvm,amazon-api-gateway"
+  echo "❌ FAILED | GRADLE  | FUNCTION  | PROVIDED | aws-lambda,aws-cdk,graalvm,amazon-api-gateway,$architecture"
   exit $EXIT_STATUS
 fi
 
@@ -131,10 +138,10 @@ rm -rf starter-cli/temp
 
 ############
 # | BUILD  | TYPE     | RUNTIME  | FEATURES
-# | GRADLE | FUNCTION | PROVIDED | aws-lambda,aws-cdk,graalvm,amazon-api-gateway-http
+# | GRADLE | FUNCTION | PROVIDED | aws-lambda,aws-cdk,graalvm,amazon-api-gateway-http,$architecture
 ############
-
-./gradlew micronaut-cli:run --args="create-function-app -b gradle -f \"aws-lambda,aws-cdk,graalvm,amazon-api-gateway-http\" temp" || EXIT_STATUS=$?
+echo "create-function-app -b gradle -f \"aws-lambda,aws-cdk,graalvm,amazon-api-gateway-http,$architecture\" temp"
+./gradlew micronaut-cli:run --args="create-function-app -b gradle -f \"aws-lambda,aws-cdk,graalvm,amazon-api-gateway-http,$architecture\" temp" || EXIT_STATUS=$?
 
 if [ $EXIT_STATUS -ne 0 ]; then
   exit $EXIT_STATUS
@@ -145,7 +152,7 @@ cd starter-cli/temp
 ./test-lambda.sh  || EXIT_STATUS=$? 
 
 if [ $EXIT_STATUS -ne 0 ]; then
-  echo "❌ FAILED | GRADLE  | FUNCTION  | PROVIDED | aws-lambda,aws-cdk,graalvm,amazon-api-gateway-http"
+  echo "❌ FAILED | GRADLE  | FUNCTION  | PROVIDED | aws-lambda,aws-cdk,graalvm,amazon-api-gateway-http,$architecture"
   exit $EXIT_STATUS
 fi
